Refactoring server.py into more intentional files
